add lombok,comment plugin
