--- generator-core/src/main/java/com/dingtalk/demo/dao/access/ChatGroupMappingMapper.java
@@ -1,96 +1,0 @@
-package com.dingtalk.demo.dao.access;
-
-import com.dingtalk.demo.domain.access.ChatGroupMapping;
-import com.dingtalk.demo.domain.access.ChatGroupMappingCriteria;
-import java.util.List;
-import org.apache.ibatis.annotations.Param;
-
-public interface ChatGroupMappingMapper {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    long countByExample(ChatGroupMappingCriteria example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    int deleteByExample(ChatGroupMappingCriteria example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    int deleteByPrimaryKey(Long id);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    int insert(ChatGroupMapping record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    int insertSelective(ChatGroupMapping record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    List<ChatGroupMapping> selectByExample(ChatGroupMappingCriteria example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    ChatGroupMapping selectByPrimaryKey(Long id);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    int updateByExampleSelective(@Param("record") ChatGroupMapping record, @Param("example") ChatGroupMappingCriteria example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    int updateByExample(@Param("record") ChatGroupMapping record, @Param("example") ChatGroupMappingCriteria example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    int updateByPrimaryKeySelective(ChatGroupMapping record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    int updateByPrimaryKey(ChatGroupMapping record);
-}
