--- generator-core/src/main/java/com/dingtalk/demo/domain/access/ChatGroupMapping.java
@@ -1,212 +1,0 @@
-package com.dingtalk.demo.domain.access;
-
-import java.io.Serializable;
-import java.util.Date;
-
-public class ChatGroupMapping implements Serializable {
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column chat_group_mapping.id
-     *
-     * @mbg.generated
-     */
-    private Long id;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column chat_group_mapping.gmt_create
-     *
-     * @mbg.generated
-     */
-    private Date gmtCreate;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column chat_group_mapping.gmt_modified
-     *
-     * @mbg.generated
-     */
-    private Date gmtModified;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column chat_group_mapping.org_id
-     *
-     * @mbg.generated
-     */
-    private Long orgId;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column chat_group_mapping.cid
-     *
-     * @mbg.generated
-     */
-    private String cid;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column chat_group_mapping.code
-     *
-     * @mbg.generated
-     */
-    private String code;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table chat_group_mapping
-     *
-     * @mbg.generated
-     */
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column chat_group_mapping.id
-     *
-     * @return the value of chat_group_mapping.id
-     *
-     * @mbg.generated
-     */
-    public Long getId() {
-        return id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column chat_group_mapping.id
-     *
-     * @param id the value for chat_group_mapping.id
-     *
-     * @mbg.generated
-     */
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column chat_group_mapping.gmt_create
-     *
-     * @return the value of chat_group_mapping.gmt_create
-     *
-     * @mbg.generated
-     */
-    public Date getGmtCreate() {
-        return gmtCreate;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column chat_group_mapping.gmt_create
-     *
-     * @param gmtCreate the value for chat_group_mapping.gmt_create
-     *
-     * @mbg.generated
-     */
-    public void setGmtCreate(Date gmtCreate) {
-        this.gmtCreate = gmtCreate;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column chat_group_mapping.gmt_modified
-     *
-     * @return the value of chat_group_mapping.gmt_modified
-     *
-     * @mbg.generated
-     */
-    public Date getGmtModified() {
-        return gmtModified;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column chat_group_mapping.gmt_modified
-     *
-     * @param gmtModified the value for chat_group_mapping.gmt_modified
-     *
-     * @mbg.generated
-     */
-    public void setGmtModified(Date gmtModified) {
-        this.gmtModified = gmtModified;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column chat_group_mapping.org_id
-     *
-     * @return the value of chat_group_mapping.org_id
-     *
-     * @mbg.generated
-     */
-    public Long getOrgId() {
-        return orgId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column chat_group_mapping.org_id
-     *
-     * @param orgId the value for chat_group_mapping.org_id
-     *
-     * @mbg.generated
-     */
-    public void setOrgId(Long orgId) {
-        this.orgId = orgId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column chat_group_mapping.cid
-     *
-     * @return the value of chat_group_mapping.cid
-     *
-     * @mbg.generated
-     */
-    public String getCid() {
-        return cid;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column chat_group_mapping.cid
-     *
-     * @param cid the value for chat_group_mapping.cid
-     *
-     * @mbg.generated
-     */
-    public void setCid(String cid) {
-        this.cid = cid == null ? null : cid.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column chat_group_mapping.code
-     *
-     * @return the value of chat_group_mapping.code
-     *
-     * @mbg.generated
-     */
-    public String getCode() {
-        return code;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column chat_group_mapping.code
-     *
-     * @param code the value for chat_group_mapping.code
-     *
-     * @mbg.generated
-     */
-    public void setCode(String code) {
-        this.code = code == null ? null : code.trim();
-    }
-}
