--- generator-core/src/main/resources/generatorConfigMyBatis3.xml
@@ -1,81 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE generatorConfiguration
-        PUBLIC "-//mybatis.org//DTD MyBatis Generator Configuration 1.0//EN"
-        "http://mybatis.org/dtd/mybatis-generator-config_1_0.dtd">
-
-<generatorConfiguration>
-
-    <context id="generatorTables" targetRuntime="MyBatis3">
-        <property name="mergeable" value="false"/>
-
-        <plugin type="org.mybatis.generator.plugins.MapperConfigPlugin">
-            <property name="fileName" value="mybatis-config.xml"/>
-            <property name="targetPackage" value="/"/>
-            <property name="targetProject" value="src/main/resources"/>
-        </plugin>
-
-        <!--  此处是将Example改名为Criteria 当然 想改成什么都行  -->
-        <plugin type="org.mybatis.generator.plugins.RenameExampleClassPlugin">
-            <property name="searchString" value="Example"/>
-            <property name="replaceString" value="Criteria"/>
-        </plugin>
-
-        <plugin type="org.mybatis.generator.plugins.SerializablePlugin"/>
-
-        <!-- commentGenerator 去除自动生成的注释  -->
-        <commentGenerator>
-            <property name="suppressDate" value="true" />
-        </commentGenerator>
-        <!--
-                <jdbcConnection driverClass="com.mysql.jdbc.Driver"
-                connectionURL="jdbc:mysql://127.0.0.1:3306/wjs?useUnicode=true&amp;characterEncoding=utf8&amp;zeroDateTimeBehavior=convertToNull&amp;remarksReporting=true" userId="root"
-                password="hundsun">
-                </jdbcConnection>
-         -->
-        <jdbcConnection driverClass="com.mysql.jdbc.Driver"
-                        connectionURL="jdbc:mysql://127.0.0.1:3306/temp_utf8"
-                        userId="root"
-                        password="root">
-        </jdbcConnection>
-        <!--
-        默认false，把JDBC DECIMAL 和 NUMERIC 类型解析为 Integer true，
-        把JDBC DECIMAL 和 NUMERIC 类型解析为java.math.BigDecimal
-        -->
-        <javaTypeResolver>
-            <property name="forceBigDecimals" value="false" />
-        </javaTypeResolver>
-
-        <!-- javaModelGenerator是模型的生成信息，这里将指定这些Java model类的生成路径； -->
-        <javaModelGenerator targetPackage="com.dingtalk.demo.domain.access"
-                            targetProject="src/main/java">
-            <property name="enableSubPackages" value="true" />
-            <property name="trimStrings" value="true" />
-        </javaModelGenerator>
-
-
-
-
-        <!-- sqlMapGenerator是mybatis 的sqlMapper XML文件的生成信息，包括生成路径等；   先生成xml，在生成java-->
-        <sqlMapGenerator targetPackage="com.dingtalk.demo.dao.access"
-                         targetProject="src/main/resources">
-            <property name="enableSubPackages" value="true" />
-        </sqlMapGenerator>
-
-        <!-- javaClientGenerator是应用接口的生成信息； -->
-        <javaClientGenerator type="XMLMAPPER"
-                             targetPackage="com.dingtalk.demo.dao.access"
-                             targetProject="src/main/java">
-            <property name="enableSubPackages" value="true" />
-        </javaClientGenerator>
-
-        <!-- table是用户指定的被生成相关信息的表，它必须在指定的jdbc连接中已经被建立。可以多个 -->
-        <table tableName="chat_group_mapping" domainObjectName="ChatGroupMapping" ></table>
-        <!--
-        <table tableName="access_custid_control_config" domainObjectName="AccessCustidControlConfig" ></table>
-        <table tableName="access_flow_control_config" domainObjectName="AccessFlowControlConfig" ></table>
-        <table tableName="access_ip_control_config" domainObjectName="AccessIpControlConfig" ></table>
-         -->
-
-
-    </context>
-</generatorConfiguration>
