--- generator-core/src/main/resources/mybatis-config.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE configuration PUBLIC "-//mybatis.org//DTD Config 3.0//EN" "http://mybatis.org/dtd/mybatis-3-config.dtd">
-<configuration>
-  <!--
-    This file is generated by MyBatis Generator.
-    This file is the shell of a Mapper Config file - in many cases you will need to add
-      to this file before it is usable by MyBatis.
-    This file was generated on Tue Sep 29 16:22:44 CST 2020.
-  -->
-  <mappers>
-    <mapper resource="com/dingtalk/demo/dao/access/ChatGroupMappingMapper.xml" />
-  </mappers>
-</configuration>
