--- generator-core/src/test/java/org/mybatis/generator/MyBatisGeneratorTest.java
@@ -1,33 +1,0 @@
-package org.mybatis.generator;
-
-import static org.junit.Assert.assertEquals;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.junit.Test;
-import org.mybatis.generator.api.MyBatisGenerator;
-import org.mybatis.generator.config.Configuration;
-import org.mybatis.generator.config.xml.ConfigurationParser;
-import org.mybatis.generator.exception.InvalidConfigurationException;
-import org.mybatis.generator.internal.DefaultShellCallback;
-
-public class MyBatisGeneratorTest {
-
-    @Test
-    public void testGenerateMyBatis3() throws Exception {
-        List<String> warnings = new ArrayList<String>();
-        ConfigurationParser cp = new ConfigurationParser(warnings);
-        Configuration config = cp.parseConfiguration(this.getClass().getClassLoader().getResourceAsStream("generatorConfigMyBatis3.xml"));
-
-        DefaultShellCallback shellCallback = new DefaultShellCallback(true);
-
-        try {
-            MyBatisGenerator myBatisGenerator = new MyBatisGenerator(config, shellCallback, warnings);
-            myBatisGenerator.generate(null);
-        } catch (InvalidConfigurationException e) {
-            assertEquals(2, e.getErrors().size());
-        }
-    }
-
-}
