+++ generator-dao/src/test/java/Test.java
@@ -1,0 +1,2 @@
+public class Test {
+}
