+++ generator-plugin/src/main/java/com/dingtalk/demo/domain/access/ChatGroupMapping.java
@@ -1,0 +1,58 @@
+package com.dingtalk.demo.domain.access;
+
+import java.io.Serializable;
+import java.util.Date;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+/**
+ * 多群活码映射表
+ * @author szy
+ */
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class ChatGroupMapping implements Serializable {
+    /**
+     * 多群码主键
+     */
+    private Long orgId;
+
+    /**
+     * 多群码
+     */
+    private String code;
+
+    /**
+     * id
+     */
+    private Long id;
+
+    /**
+     * 创建时间
+     */
+    private Date gmtCreate;
+
+    /**
+     * 修改时间
+     */
+    private Date gmtModified;
+
+    /**
+     * 会话ID
+     */
+    private String cid;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database table chat_group_mapping
+     *
+     * @mbg.generated Tue Oct 13 19:18:22 CST 2020
+     */
+    private static final long serialVersionUID = 1L;
+}
