+++ generator-plugin/src/main/java/com/dingtalk/generator/plugins/SelectByUniqueKeyPlugin.java
@@ -1,0 +1,8 @@
+package com.dingtalk.generator.plugins;
+
+/**
+ * select by uniqueKey, should assign unique key name
+ * @author suyin
+ */
+public class SelectByUniqueKeyPlugin {
+}
