+++ generator-plugin/src/main/java/com/dingtalk/generator/plugins/UpdateByUniqueKeySelectivePlugin.java
@@ -1,0 +1,4 @@
+package com.dingtalk.generator.plugins;
+
+public class UpdateByUniqueKeySelectivePlugin {
+}
